chore: rename `main-v2.py` to `main.py`
