+++ docker-compose.yml
@@ -7,7 +7,7 @@ services:
       - "8000:8000"
     volumes:
       - ./server:/app
-    command: ["uvicorn", "main-v2:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
+    command: ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
 
   client:
     build:
